103818 Add boolean parameter to ItemServiceImpl#inheritCollectionDefaultPolicies to decide whether to override item read policies

+++ dspace-api/src/main/java/org/dspace/content/ItemServiceImpl.java
@@ -920,12 +920,21 @@ public class ItemServiceImpl extends DSpaceObjectServiceImpl<Item> implements It
     @Override
     public void inheritCollectionDefaultPolicies(Context context, Item item, Collection collection)
         throws SQLException, AuthorizeException {
+        inheritCollectionDefaultPolicies(context, item, collection, true);
+    }
+
+    @Override
+    public void inheritCollectionDefaultPolicies(Context context, Item item, Collection collection,
+                                                 boolean overrideItemReadPolicies)
+        throws SQLException, AuthorizeException {
 
         // If collection has READ policies, remove the item's READ policies.
-        List<ResourcePolicy> defaultCollectionPolicies = authorizeService
-            .getPoliciesActionFilter(context, collection, Constants.DEFAULT_ITEM_READ);
-        if (!defaultCollectionPolicies.isEmpty()) {
-            authorizeService.removePoliciesActionFilter(context, item, Constants.READ);
+        if (overrideItemReadPolicies) {
+            List<ResourcePolicy> defaultCollectionPolicies = authorizeService
+                .getPoliciesActionFilter(context, collection, Constants.DEFAULT_ITEM_READ);
+            if (!defaultCollectionPolicies.isEmpty()) {
+                authorizeService.removePoliciesActionFilter(context, item, Constants.READ);
+            }
         }
 
         adjustItemPolicies(context, item, collection);
